# swagger.json
{"openapi":"3.1.0","info":{"title":"Bob Family Tree API","description":"Open /docs to test everything!","version":"1.0.0"},"paths":{"/persons":{"get":{"tags":["Person"],"summary":"List All Persons","description":"Returns ALL persons in the database with their firstnames, lastnames, DOB and DOD","operationId":"List_all_persons_persons_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["Person"],"summary":"Add A New Person To The Database","description":"Creates a new person. At least one firstname is required and last name, DOB and DOD are optional","operationId":"Add_a_new_person_to_the_database_persons_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}":{"get":{"tags":["Person"],"summary":"Get Information About A Specific Person","description":"Returns ONE specific person by their ID.","operationId":"Get_information_about_a_specific_person_persons__person_id__get","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Person"],"summary":"Add Or Update Birth Date And Death Date Of A Person","description":"Adds or updates the birth date and/or death date of an existing person.\nOnly send the field you want to change — the other one will stay the same.\nSend null to remove a date.","operationId":"Add_or_Update_birth_date_and_death_date_of_a_person_persons__person_id__patch","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Person"],"summary":"Delete A Specific Person","description":"Deletes a person permanently.\nTheir names and relationships will be deleted automatically as well (CASCADE).","operationId":"Delete_a_specific_person_persons__person_id__delete","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/firstname":{"post":{"tags":["FirstName"],"summary":"Add A New Firstname To A Person","description":"This will create only a firstname for a specific person using the person_id","operationId":"Add_a_new_firstname_to_a_person_persons__person_id__firstname_post","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FirstnameBody"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/firstnames/{firstname_id}":{"put":{"tags":["FirstName"],"summary":"Update A Firstname","description":"Rename an existing firstname using the person_id and firstname_id","operationId":"Update_a_firstname_persons__person_id__firstnames__firstname_id__put","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}},{"name":"firstname_id","in":"path","required":true,"schema":{"type":"integer","title":"Firstname Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FirstnameBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/persons/{person_id}/firstnames/{firstname_id}":{"delete":{"tags":["FirstName"],"summary":"Delete Firstname","description":"Delete the firstname from a person using the person_id and firstname_id","operationId":"Delete_firstname_delete_persons__person_id__firstnames__firstname_id__delete","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}},{"name":"firstname_id","in":"path","required":true,"schema":{"type":"integer","title":"Firstname Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/lastname":{"post":{"tags":["LastName"],"summary":"Add A New Lastname To A Person","description":"This will create only a lastname for a specific person using the person_id","operationId":"Add_a_new_lastname_to_a_person_persons__person_id__lastname_post","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LastnameBody"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/lastnames/{lastname_id}":{"put":{"tags":["LastName"],"summary":"Update A Lastname","description":"Rename an existing lastname using the person_id and lastname_id","operationId":"Update_a_lastname_persons__person_id__lastnames__lastname_id__put","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}},{"name":"lastname_id","in":"path","required":true,"schema":{"type":"integer","title":"Lastname Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LastnameBody"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["LastName"],"summary":"Delete Lastname","description":"Delete specific lastname using the person_id and lastname_id","operationId":"Delete_lastname_persons__person_id__lastnames__lastname_id__delete","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}},{"name":"lastname_id","in":"path","required":true,"schema":{"type":"integer","title":"Lastname Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/parents":{"post":{"tags":["Relationships"],"summary":"Add Parent To A Person","description":"This will create only a parent for a specific person using the person_id","operationId":"Add_parent_to_a_person_persons__person_id__parents_post","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ParentAdd"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Relationships"],"summary":"Get All Parents Of Person","description":"Returns all parents of a person using the person_id.","operationId":"Get_all_parents_of_person_persons__person_id__parents_get","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/children":{"get":{"tags":["Relationships"],"summary":"Get All Children Of Person","description":"Returns all children of a person.","operationId":"Get_all_children_of_person_persons__person_id__children_get","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/persons/{person_id}/parents/{parent_id}":{"delete":{"tags":["Relationships"],"summary":"Remove Parent","description":"Removes the relationship between a parent and a child.","operationId":"Remove_parent_persons__person_id__parents__parent_id__delete","parameters":[{"name":"person_id","in":"path","required":true,"schema":{"type":"integer","title":"Person Id"}},{"name":"parent_id","in":"path","required":true,"schema":{"type":"integer","title":"Parent Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"FirstnameBody":{"properties":{"firstname":{"type":"string","title":"Firstname"}},"type":"object","required":["firstname"],"title":"FirstnameBody"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LastnameBody":{"properties":{"lastname":{"type":"string","title":"Lastname"}},"type":"object","required":["lastname"],"title":"LastnameBody"},"ParentAdd":{"properties":{"parent_id":{"type":"integer","title":"Parent Id"},"relationship_type":{"type":"string","title":"Relationship Type","default":"biological"}},"type":"object","required":["parent_id"],"title":"ParentAdd"},"PersonCreate":{"properties":{"firstnames":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Firstnames","default":[]},"lastnames":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Lastnames","default":[]},"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"},"death_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Death Date"}},"type":"object","title":"PersonCreate"},"PersonUpdate":{"properties":{"birth_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Birth Date"},"death_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Death Date"}},"type":"object","title":"PersonUpdate"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"tags":[{"name":"Person","description":"Work with persons in the family tree."},{"name":"Names","description":"Add, rename or delete firstnames and lastnames."},{"name":"Relationships","description":"Manage parent and child links between persons."},{"name":"FirstName","description":"Add, rename or delete firstnames of a person."},{"name":"LastName","description":"Add, rename or delete lastnames of a person."},{"name":"Relationships","description":"Manage parent and child links between persons."}]}